Move test to test_config_flow
